requirements.txt
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.6,<2.10
drf-spectacular>=0.26.2,<0.27
Pillow>=9.5.0,<9.6
django-money>=3.1.0,<3.2.0
geopy>=2.3.0,<2.4.0
#uwsgi>=2.0.21,<2.1